Test to have angular working properly

+++ backend/src/main/kotlin/com/techso/techsport/model/response/PagesResponse.kt
@@ -1,3 +1,3 @@
 package com.techso.techsport.model.response
 
-data class PagesResponse(var pages: Long)
+data class PagesResponse(var pages: Int)
